fixed incomplete ALSA9 support and changed sounds.dist to use .raw
The ALSA9 support was in configure but the code part wasn't in the cfsndserv.c
Applied the patch from Tim Hentenaar submitted on mailing list in march 2003 to
correct this.
Also changed the default sound file so it uses .raw files and not .au because
CVS sounds are now .raw.


git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/client@2541 282e977c-c81d-0410-88c4-b93c2d0d6712