Building disappear bug fix
Fixed a bug of buildings disappearing when moving from west to east or
north to south in cities where building are made of a bigimage,
covering multisquares. This was because every square entering the map
view after a scroll was forced to empty, waiting for info from server.
But bigimage buildings are sended by server before they enter map. Added the flag
to ask no reset of the bigimage during the erasing of squares and setting flag to
true for squares entering the map.


git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/client@2542 282e977c-c81d-0410-88c4-b93c2d0d6712