Correctly set maxfd
The nfds argument to select() is the number of the highest file
descriptor that we're interested in. sysconf(_SC_OPEN_MAX) returns *one
more* than the largest file descriptor allocated by the system, which
causes select() to fail on FreeBSD. One fix would be to subtract one
from this number, but it is better to just make  maxfd a local variable
set to the highest script file descriptor.

While here, unify the exec failure error message and print errors in the
future if select() fails.

git-svn-id: svn://svn.code.sf.net/p/crossfire/code/client/trunk@21720 282e977c-c81d-0410-88c4-b93c2d0d6712