Add local map scroll prediction
Latency to a Crossfire server can be significant, making even basic
movement unpleasant across a network. Player characters with low speeds
experience a timeout after every move that feels like network latency.
Local map scroll prediction gives player instant feedback when they
issue movement commands by shifting the offset of the rendered map on
screen. This gives the illusion that the Crossfire server is responding
faster than it actually is.

Local map scroll prediction doesn't know whether a movement command
succeeded or not, only that it has been acknowledged by the server (via
the 'comc' command). But it doesn't actually care: regardless of whether
movement was successful or not (e.g. player ran into a wall), the player
should always be rendered in the middle of the map. The strategy used by
this code is to predict the movement, then cancel the prediction as soon
as the movement command is acknowledged. If no move takes place, the
animation snaps back, as if the player bounced off the wall.

To make prediction work across more than one movement command, the
client now stores a ring buffer of the directions of the last 256
commands. Several movement commands can be queued at once and the client
will show the correct prediction.