Render fog of war map beyond server viewport
If the map drawing area is larger than the biggest viewport the server
can support, the map is clipped at the server viewport boundaries
regardless of the drawing area size and fog of war settings.

This change allows the client to draw beyond the server viewport using
its fog of war data. In addition, the client now keeps the player
centered in the drawing area, even at mapscale < 100.