Add initial sandboxing support
Add a checkbox to enable sandboxing to the metaserver window. When
sandboxing is enabled on platforms that support it (currently only
FreeBSD), the client drops into a sandbox with limited permissions right
after connecting to the server. This limits the impact of security
vulnerabilities in the client or dependent libraries when connecting to
untrusted servers.

Currently this is still somewhat broken in the sense that GTK themes and
fonts are lazy-loaded, which means that themes and some fonts do not
show up correctly when sandboxing is enabled. One temporary workaround
for this issue is to first connect to a trusted server so that GTK can
cache themes and fonts, disconnect, and then enable sandboxing when
connecting to an untrusted server.